Added more Storybook stories and pulled out SnackbarItem into its own file.
