+++ app/javascript/Snackbar/index.js
@@ -1,0 +1,2 @@
+export * from './Snackbar';
+export * from './SnackbarItem';
